Named Fixes and BMC Network Config
Used semicolons instead of hashes for commenting in the named templates
to prevent issues with named running (it doesn't seem to interpret
hashes as comments despite what documentation says).

Correct the rescues of asset OOB info such that not all nodes have to
have linked assets in order for any rendering to work.